move db types to types folder
